proc: Remove verifiedbootstate flag from /proc/cmdline
From: Sultanxda <sultanxda@gmail.com>
Date: Fri, 21 Oct 2016 01:37:16 -0700

Userspace parses this and sets the ro.boot.verifiedbootstate prop
according to the value that this flag has. When ro.boot.verifiedbootstate
is not 'green', SafetyNet is tripped and fails the CTS test.

Hide verifiedbootstate from /proc/cmdline in order to fix the failed
SafetyNet CTS check.