Make lgecrypto_module.ko loadable in new built kernel
Due to a previous mistake, lgecrypto_module.ko was copied to /system/lib/modules
and this is a wrong behavior. Fixing this by revert lgecrypto_module.ko back to
where it was and make it loadable by new built kernel. Althrough it's loadable but
we should not load it for now because it will cause kernel panic.